move component models to new folder structure WIP
