+++ examples/bri-3/src/bri/identity/bpiSubjects/models/bpiSubject.interface.ts
