+++ examples/bri-3/src/bri/identity/bpiSubjects/models/bpiSubjectType.enum.ts
