+++ examples/bri-3/src/bri/workgroup/invitations/models/invitation.type.ts
