+++ examples/bri-3/src/bri/workgroup/workflows/models/workflow.ts
