+++ examples/bri-3/src/bri/workgroup/workgroups/models/workgroup.ts
