+++ examples/bri-3/src/bri/workgroup/workgroups/models/workgroupComponent.interface.ts
