+++ examples/bri-3/src/bri/workgroup/worksteps/models/workstep.interface.ts
@@ -1,4 +1,4 @@
-import { Privacy } from '../../policy/privacy';
+import { Privacy } from '../../ policy/privacy';
 import { Security } from '../../policy/security';
 
 export interface IWorkstep {
