+++ examples/bri-3/src/bri/workgroup/worksteps/models/workstep.ts
